Refactor and update admin features
